% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{aggregate_classic}
\alias{aggregate_classic}
\title{Calculates the average anothter neightourhood test}
\usage{
aggregate_classic(dat_nb)
}
\arguments{
\item{dat_nb}{a neightbourhood table with the labels applied with 'apply_labels'}
}
\description{
Specifically: How many neightbours of type B does a cell of type A have.
}
