% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{aggregate_classic_patch}
\alias{aggregate_classic_patch}
\title{Calculates patch detection neightbourhood statistics}
\usage{
aggregate_classic_patch(dat_nb, patch_size)
}
\arguments{
\item{dat_nb}{a neightbourhood table with the labels applied with 'apply_labels'}

\item{patch_size}{minimum number of neighbours to be considered a patch}
}
\value{
a statistics with columns group, Firstlabel, Secondlabel, ct
}
\description{
Calculates: What fraction of cells of type A have at least a given number of neighbours of type B?
}
