% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{aggregate_classic_wrong}
\alias{aggregate_classic_wrong}
\title{Calculate yet another neightourhood statistics
This calculates 'what are the average number of interactions between two labels}
\usage{
aggregate_classic_wrong(dat_nb)
}
\description{
Calculate yet another neightourhood statistics
This calculates 'what are the average number of interactions between two labels
}
