% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{apply_labels}
\alias{apply_labels}
\title{Applies the labels on a a relationship table}
\usage{
apply_labels(dat_labels, dat_rel)
}
\arguments{
\item{dat_labels}{a labels table as formated by the prepare_tables function. Must have columns 'label', 'ObjectID', 'group'}

\item{dat_rel}{a relationship table as formated by the prepare_tables function. Must have columns "First Object ID", "Second Object ID"}
}
\value{
a copy of the dat_rel table with additional coluns FirstLabel and SecondLabel
}
\description{
Applies the labels on a a relationship table
}
