% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood-package.R
\docType{package}
\name{neighbouRhood-package}
\alias{neighbouRhood}
\alias{neighbouRhood-package}
\title{neighbouRhood: neighbouRhood analysis in R}
\description{
Reproduces the neightbourhood analysis from the HistoCAT paper in R
}
\keyword{internal}
