% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{prepare_tables}
\alias{prepare_tables}
\title{prepare_tables
This data prepares CellProfiller style input tables into standardized tables
optimized for the neightbourhood analysis.}
\usage{
prepare_tables(
  dat_obj,
  dat_rel,
  objname = DEFAULTOBJNAME,
  relationship = DEFAULTRELATIONSHIP,
  col_imnr = IMGNR,
  col_objnr = OBJNR,
  col_label = LABEL,
  col_group = NULL,
  col_relationship = RELATIONSHIP,
  col_objname = OBJNAME
)
}
\arguments{
\item{dat_obj}{an object level data.table from cellprofiler}

\item{dat_rel}{a relationship table from cellprofiler}

\item{objname}{the objectname of the object table}

\item{relationship}{the relationship table}

\item{col_imnr}{the image number column}

\item{col_objnr}{the object number column}

\item{col_label}{the label/cluster number to do the permutation test on}

\item{col_group}{a grouping to do the permutation test with, default: do it per image}

\item{col_relationship}{the column name for the relationship column}

\item{col_objname}{the column for the object name column}
}
\value{
returns the percentile of each value of x
     a list with elements:

          [[1]] a obj table with columns:
             'ObjectID': a new, integer ID from 1:N
             'label': the group/cluster label of the cells
             'ImageNumber': the CP imagenumber
             'ObjectNumber': the CP objectnumber
             'Group': The group to do the permutations by
             'ObjectName': The object name

          [[2]] a relationship table with columns:
           'First Object ID': Id of the first object
           'Second Object ID': Id of the second object
           'Group': the group to which the relationship belongs
}
\description{
prepare_tables
This data prepares CellProfiller style input tables into standardized tables
optimized for the neightbourhood analysis.
}
