% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{shuffle_labels}
\alias{shuffle_labels}
\title{shuffle_labels}
\usage{
shuffle_labels(dat_labels)
}
\arguments{
\item{dat_labels}{a labels table as formated by the prepare_tables function. Must have columns 'label', 'ObjectID', 'group'}
}
\value{
returns a copy of dat_labels with shuffeled labels within the group
}
\description{
This shuffles the labels by group in a standardized dat_labels table
}
