% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{aggregate_classic}
\alias{aggregate_classic}
\title{Calculates the the `average neightourhood`}
\usage{
aggregate_classic(dat_nb)
}
\arguments{
\item{dat_nb}{a neightbourhood table with the labels applied with 'apply_labels'}
}
\value{
a statistics with columns group, Firstlabel, Secondlabel, ct
}
\description{
Calculates: How many neightbours of type B does a cell of type A have on average.
}
