% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouRhood.R
\name{aggregate_histo}
\alias{aggregate_histo}
\title{Calculates HistoCAT style permutation test}
\usage{
aggregate_histo(dat_nb)
}
\arguments{
\item{dat_nb}{a neightbourhood table with the labels applied with 'apply_labels'}
}
\description{
Calculates: How many many neightbours of type B has a cell of type A given it has at least one neigthbour of type B?
}
